wsd: move wopi stub/dummy interface to MobileApp.hpp
