+++ wsd/DocumentBroker.hpp
@@ -37,34 +37,9 @@
 #if !MOBILEAPP
 #include "Admin.hpp"
 #include <wopi/WopiStorage.hpp>
-#else
-//FIXME: Move to a stub file.
-class WopiStorage : public StorageBase
-{
-public:
-    class WOPIFileInfo : public FileInfo
-    {
-    public:
-        enum class TriState
-        {
-            False,
-            True,
-            Unset
-        };
-
-        std::string getTemplateSource() const { return std::string(); }
-
-        bool getDisablePrint() const { return false; }
-        bool getDisableExport() const { return false; }
-        bool getDisableCopy() const { return false; }
-        bool getEnableOwnerTermination() const { return false; }
-
-        TriState getDisableChangeTrackingShow() const { return TriState::Unset; }
-        TriState getDisableChangeTrackingRecord() const { return TriState::Unset; }
-        TriState getHideChangeTrackingControls() const { return TriState::Unset; }
-    };
-};
-#endif
+#else // MOBILEAPP
+#include <MobileApp.hpp>
+#endif // MOBILEAPP
 
 // Forwards.
 class PrisonerRequestDispatcher;
